+++ docs/zh/guide/changelog.md
@@ -6,7 +6,7 @@
 
 #### 1.2.10
 
-- feat: 新增 `inline-script` 内联脚本功能，支持在 Markdown 中使用 `vp-script` 编写当前页面专用脚本。
+- feat: 新增 `vp-script` 内联脚本功能，支持在 Markdown 中使用 `vp-script` 编写当前页面专用脚本。
 - feat: 新增 `footer-script` 页脚脚本功能，支持通过 `vp/config/footerScript.ts` 为所有页面统一注入底部脚本。
 - chore: 更新依赖版本
 - fix: 根据依赖更新，收窄类型，让类型定义更严格和精准
@@ -51,3 +51,12 @@
 
 - fix: 完成 `create-vanilla-press` 和 `vanilla-press` 的架构重构。
 - style: 修复侧边栏滚动条问题
+
+#### 1.3.1
+
+- docs: 新增 repo README
+
+#### 1.3.2
+
+- fix: seo 首页标题规则
+- docs: 更新布局/组件/vp-script文档。